+++ src/pomme/maconv/stuffit/methods.h
@@ -45,7 +45,7 @@ struct CompressionMethod {
 
 
     // Read the next bytes.
-    virtual int32_t ReadBytes(uint8_t *data, uint32_t length) { return -1; }
+    virtual int32_t ReadBytes(uint8_t *data, size_t length) { return -1; }
 
     // Initialize the algorithm.
     virtual void Initialize() {}
